refactor: Organize code into folders
