+++ src/tree/AppMapTreeDataProvider.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import AppMapCollection from '../appmapCollection';
+import AppMapCollection from '../services/appmapCollection';
 
 const LABEL_NO_NAME = 'Untitled AppMap';
 export class AppMapTreeDataProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
