+++ src/tree/index.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import AppMapCollectionFile from '../appmapCollectionFile';
+import AppMapCollectionFile from '../services/appmapCollectionFile';
 import Links from './links';
 import { InstructionsTreeDataProvider } from './instructionsTreeDataProvider';
 import { AppMapTreeDataProvider } from './AppMapTreeDataProvider';
